Revamp image-throwing animation in MainActivity
The image-throwing animation is now a more accurate physics simulation.
Gravity is set at a fixed value (@dimen/thrown_image_gravity), the
horizontal speed of the image is generated, and the maximum height,
difference between the start and end vertical positions, and total time
is found. This is used to position the image on-screen, and the throwing
itself takes place in accordance with kinematics. Ranges can all be
found in dimens.xml.

Note that in order to provide a more consistent animation across all
devices, gravity and the range of horizontal speeds are scaled based on
screen size. Gravity, @dimen/thrown_image_gravity = -700 dp/s/s, is
based on a screen with height = 1280 dp, so the gravity value is
multiplied by (height / 1280). Similarly, the range of horizontal
speeds (450-700 dp/s) is based on a screen with width = 800 dp, so the
horizontal speed is multiplied by (width / 800).