Adjust main animation in landscape on tablets
When width was greater than height, the images in the image-throwing
animation were previously flying across the screen. This has been fixed:
when width is greater than height, the two are swapped for the purpose
of adjusting gravity and horizontal speed.

Also, the difficulty buttons are now shorter on tablets in landscape
mode for less overlap with the image-throwing animation.