Make AboutActivity look good on tablets
On tablets, the text in AboutActivity is now larger. On all devices,
the text in AboutActivity is now vertically centred; this required
changing the root layout to ConstraintLayout.

As well, many properties are shared between the sign in/out button in
AboutActivity and the difficulty buttons. Some of these properties now
use shared dimensions via @dimen/full_length_buttons_text_size and
@dimen/full_length_buttons_text_padding. The sign in/out button's
side margins were also adjusted for tablet dimensions.