Format pause overlay for tablets in landscape mode
The pause overlay now looks good for both sw600dp and sw720dp tablets in
landscape mode.

Note that in order to move the elements of the pause screen to the side,
the pause overlay actually only takes up the right portion of the
screen. To counter this, the background of the screen itself is now
changed in GameActivity.

As well, the new layout-land/game_paused.xml is nearly identical to
layout/game_paused.xml; this redundancy should probably be eliminated
through combining the files somehow.