Format PostGameActivity for tablets in portrait
PostGameActivity now looks good on both sw600dp and sw720dp tablets in
portrait mode.

Note that to accomplish this, PostGameActivity's root layout was changed
from RelativeLayout to ConstraintLayout. This allowed for a more
flexible design, as well as finally fixing a bug in which the
achievement and leaderboard buttons' text was misaligned.