Format PostGameActivity for sw720dp in landscape
PostGameActivity now has a dedicated layout for sw720dp (large) tablets
in landscape mode. Notably, it does not feature the achievement and
leaderboard buttons, instead opting for plenty of whitespace.

PostGameActivity itself also had to be changed so as not to error when
the achievement and leaderboard buttons are not in the layout.