Fix NPE in PostGameActivity.onResume()
When the "New Best Score!" pulse animation was playing and the game was
exited and re-entered such that onResume() was called, PostGameActivity
would crash with a NullPointerException. This stemmed from the fact that
the pulse animation apparently did not save across the pause/resume
cycle, and thus was null. Fixed by calling updateUi() before starting
the animation, which sets the "New Best Score!" pulse animation.

Also, fixed a warning in PostGameActivity.onCreate() by checking for the
case where getIntent().getExtras() == null. In that case, onCreate()
will now throw an appropriate IllegalArgumentException instead of a
cryptic NullPointerException.

Resolves #38.