Add basic ad removal infrastructure
AdRemovalManager is a basic utility class that statically manages
whether or not ads are removed. GameActivity (the only activity that
currently contains ads) checks AdRemovalManager.areAdsRemoved() on
startup; if that is true, it does not load the ads. Additionally, the
RemoveAdsCircleButton itself checks AdRemovalManager.areAdsRemoved() in
onAttachedToWindow(); if it is true, the circle button sets itself and
its container in the view hierarchy GONE. When the RemoveAdsCircleButton
is clicked, it calls AdRemovalManager.setAdsRemoved() to set ads as
removed.

GameActivity also contains a RemoveAdsCircleButton, so some method is
needed to dynamically remove the ads in GameActivity when that circle
button is clicked. This is done by having GameActivity implement the new
interface HasRemovableAds, which has just one method: removeAds(). When
RemoveAdsCircleButton is clicked, it checks if its context is a
HasRemovableAds, in which case the removeAds() method is called.

Currently, the layout position of items in GameActivity is not affected
by the ads' removal status, leading to large whitespace gaps when ads
are removed. This will be fixed in an upcoming commit.