Adjust GameActivity spacing with ads removed
When ads were removed, there was previously an empty space at the bottom
of GameActivity where the banner ad used to be. Now the ad is set to
GONE by default, set to VISIBLE upon loading if ads are not removed, and
set back to GONE if ads are removed in GameActivity. As well, the
GameActivity landscape layout has been adjusted so as not to look stupid
with ads removed.