Collect ad personalization consent from EEA users
The GDPR and other related laws requires apps to collect consent from
users to collect personal data such as that used by AdMob for targeted
ads. Google's Consent SDK is now used to collect and store that consent.

The default Google-rendered consent form is not translatable nor easily
customizable. For that reason, a custom consent form based on lenooh's
StackOverflow answer at https://stackoverflow.com/a/50556255 is used.
The consent form is represented by EUConsentForm, has its layout at
@layout/eu_consent_dialog (and another layout at @layout/eu_consent_
more_info_dialog), and contains a bunch of pseudo-legalese that needs to
be looked at by a lawyer-type.

The ability to modify consent in AboutActivity still needs to be
implemented - currently we aren't fully GDPR-compliant as there is no
way to modify consent.