Add ability to modify consent in AboutActivity
In AboutActivity, when the user is in the EEA and ads have not been
removed, there are now two TextViews: @id/eu_consent_status, which
explains whether the user has or has not consented, and @id/change_eu_
consent, which is a link that when clicked, pops up a cancelable
EUConsentForm where the user can change their consent status.

The GDPR has now been fully implemented (I believe).