Make EU consent form look better on phones
The dimensions and such for the EUConsentForm dialog have been adjusted
to look better on phones. A new button style (ButtonStyle.Dialog) has
been created for the consent buttons in the dialog with customizations
to make the dialog look better on small screens, and a bug with adding
the list of ad providers to the "more info" dialog has been fixed.