Update GPGSHelperService et al to use current APIs
GPGSHelperService and associated classes were using outdated APIs to
access Google Play Games Services. When the classes were written, the
GoogleApiClient-based GPGS API was current; between then and now, it was
deprecated and the new, simpler GoogleSignInAccount-based API took its
place. GPGSHelperService, GPGSAction, Score, ScoreCache, etc. now use
the GoogleSignInAccount-based API.

As well, using a List for storing actions on sign-in in
GPGSHelperService was too prone to adding actions multiple times. Now,
each activity is mapped to a 2-element array of GPGSActions. The 0th
element is the "primary" action on sign-in, which may be modified by
activities; the 1st is the "secondary" action on sign-in, which is
modified internally by GPGSHelperService as part of
tryActionOrConnect(). Additionally, activities using actions on sign-in
must now call GPGSHelperService.clearActionOnSignIn() to prevent actions
on sign-in from being called multiple times.

Resolves #43.