Fix OutOfMemoryError from loading too many images
Previously, ImageSupplier loaded all 70 512x512 PNGs at once and hung
onto them in memory for the duration of the run. Unsurprisingly, this
used a ridiculous amount of memory, and on devices with lower amounts of
memory, the app crashed immediately.

Now, ImageSupplier has been changed to load only a certain number of
images at once. When the ImageSupplier singleton is created, it loads
only the filenames of each image. To preload images, call ImageSupplier
.preload(int upTo), where upTo is the number of images that will be
loaded after the preloading is complete. Callers may also optionally
specify image names to be unloaded after preloading is complete. Other
methods are also provided. Currently, in both GameActivity and
MainActivity, @integer/images_to_preload = 15 images are always
preloaded.

In the worst-case scenario that preloading an image takes a ridiculously
long time, ImageSupplier will not run out of images; instead, the game
will be restricted only to the pool of preloaded images until loading of
more is complete.

Resolves #45.