Fix bugs in ImageSupplier
Despite the previous commit, on devices with *very* restricted memory
(such as some emulators), ImageSupplier was still running out of memory.
@integer/images_to_preload has been reduced to 10, and ImageSupplier now
handles OutOfMemoryErrors. If one occurs, it unloads any images in
toUnloadAfter, then tries again; if an OutOfMemoryError still occurs, it
skips preloading that image with an error.

As well, there was a bug in which if preload() was called while images
were still preloading, ImageSupplier would load too many images and then
would not be able to load any more. This was fixed by taking
currentlyPreloading.size() into account when calculating how many images
to load; as well, a case has been added for the case in which too many
images were loaded in order to prevent a situation like the above: if
too many images are loaded, ImageSupplier unloads random images until
that is not the case, then tries again. Infinite recursion *shouldn't*
occur.