Make close button more distinct from options
The play/close button used to be the same colour as the difficulty
option buttons, which led to confusing it for a difficulty. Now, when
clicked, the play button's colour transitions darker, making it darker
than the other buttons and less likely to be confused for a difficulty.

This is done by using a TransitionDrawable, @drawable/play_button_
transition, which transitions between @drawable/full_length_button
(which is @drawable/button, renamed) and @drawable/close_button (a new
drawable which is identical to @drawable/full_length_button except the
background colour is @color/closeButton). The transition takes @integer/
play_button_transition_ms = 200 ms.

As well, @style/ButtonStyle sounded stupid, and has been renamed to
@style/FullLengthButton.

Resolves #47.