Resize mouse pointer variant to mesh with normal
The mouse pointer variant (the fist mouse) was previously a lot larger
than the normal mouse pointer, and it was very easy to tell the
difference. The variant has been resized to be roughly the same size as
the normal without the 'fingers', making this particular variant more
difficult (marginally).