Make the sign-in/out button's text resize to fit
Previously, the English "Sign in to/out from Google Play" *just* fit in
one line in the sign-in/out button on narrower screens, but the French
"Connectez-vous à/Déconnectez-vous de Google Play" did not. Now using a
similar technique to the resizing of the paused text from the last
commit, the text resizes to fit the button in one line.

Resolves #50.