Fix sign in button not changing when signing in
The "Sign in to Google Play" button was not changing to "Sign out of
Google Play" when signing in. As this occurs in the gpgsCallback()
method, presumably the achievement for looking at the about screen was
not being awarded when signing in from AboutActivity either.

This was fixed by resetting the actionOnSignIn to CallCallbacks just
before calling gpgsHelper.signIn().

Resolves #51.