Add link to APK download in README
There is now a link to the APK download (copperium-studios.github.com/
variant-tap.apk), as well as a QR code pointing to the same link, in
README.md.

As well, the various Gradle build files have been updated to place
jcenter() at the bottom of the list of dependency sources in order to
resolve a strange build error, and a final change in dimens.xml is
included.