Migrate to AndroidX, update libs, target API 29
A few things have happened in the Android world since the last
commit, so update every library version and the Gradle version to
the latest. We now target API level 29 (Android 10/Q) and it doesn't
seem to have broken anything, so that's good.

Apparently at some point in 2018 all the various Android support
libraries got replaced with AndroidX, so we use that now.

In mid-2018, Google discontinued GPGS' requests, quests, and gifts
APIs, and GameHelper from BaseGameUtils used requests. We didn't use
that, so remove that and hopefully everything's fine.

Finally, pro tip: make sure ~/.android/debug.keystore (and the
release keystore of course) is the same on all workstations,
otherwise everything is mostly fine, but you get a nasty bug where
GPGS sign-in just silently fails with status SIGN_IN_REQUIRED. Cool.