Fix "New Best Score!" text overlapping with score
This occurred because aligning the new-best-score text to the bottom
of the score would cause too big a gap, so previously, we aligned it to
the top of the score and added a big margin. This obviously failed on
other devices and caused overlapping with the score, so now we align
the baselines and shift down a bit.